Fix failing YourJobTest
